+++ Blog-Post/index.html
@@ -35,7 +35,7 @@
       crossorigin="anonymous"
     ></script>
     
-    <script src="./src/scripts/utils.js"></script>
+    <script  type="module" src="./src/scripts/utils/utils.js"></script>
     <script type="module" src="./src/scripts/login/main.js"></script>
     <script type="module" src="./src/scripts/login/loginHandler.js"></script>
   </body>
