+++ Blog-Post/src/pages/home.html
@@ -34,7 +34,7 @@
       integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
       crossorigin="anonymous"
     ></script>
-    <script src="../scripts/utils.js"></script>
+    <script type="module" src="../scripts/utils/utils.js"></script>
     <script type="module" src="../scripts/home/home.js"></script>
   </body>
 </html>
