+++ Blog-Post/src/sass/style.sass
@@ -111,6 +111,8 @@ form
     box-sizing: border-box
     box-shadow: 0 0 10px $boxShedow
     flex: 1
+    max-width: 1000px
+    overflow: hidden
     // min-height: 300px
     margin-bottom: 10px
     img
@@ -130,10 +132,16 @@ form
         flex-wrap: wrap
         color: $color-title
         text-transform: capitalize
+        white-space: nowrap
+        overflow: hidden
+        text-overflow: ellipsis
         .card-title
             font-size: 2rem
             font-weight: bold
             text-transform: capitalize
+            white-space: nowrap
+            overflow: hidden
+            text-overflow: ellipsis
 
 .footer
     margin-top: 10px
