+++ Blog-Post/src/scripts/header/createHeader.js
@@ -1,3 +1,6 @@
+import UI from '../utils/utils.js';
+import Storage from '../../data/storage.js';
+
 function createHeader(links, bool = false) {
   let defaultBtns = [
     {
@@ -15,7 +18,7 @@ function createHeader(links, bool = false) {
   if (bool) defaultBtns = links;
 
   let btns = [];
-  if (sessionStorage.getItem('user')) {
+  if (Storage.hasUser()) {
     btns = [
       UI.createElement('img', {
         src: 'https://mdbcdn.b-cdn.net/img/new/avatars/2.webp',
