+++ Blog-Post/src/scripts/header/headerBtnHandlers.js
@@ -1,4 +1,4 @@
-import { createNotification } from "../notification/createNotification.js";
+import Storage from '../../data/storage.js';
 const toggleThem = () => {
   const toggleBtn = document.querySelector('.toggle-theme');
   const themeIcon = toggleBtn.querySelector('i');
@@ -15,18 +15,7 @@ const logOutHandler = () => {
   if (!loginBtn) return;
 
   loginBtn.addEventListener('click', () => {
-    sessionStorage.removeItem('user');
-    if (window.location.href.includes('createPost')) {
-      createNotification(
-        'info',
-        'You have successfully logged out and will be redirected to the login page.',
-      );
-      setTimeout(() => {
-        window.location.href = '../../index.html';
-      }, 2000);
-    } else {
-      location.reload();
-    }
+    Storage.clearUserData();
   });
 };
