+++ Blog-Post/src/scripts/home/newPost/createPostHandler.js
@@ -1,3 +1,4 @@
+import Storage from '../../../data/storage.js';
 import { createNotification } from '../../notification/createNotification.js';
 
 export const createNewPost = () => {
@@ -6,7 +7,7 @@ export const createNewPost = () => {
 
   addBnt.addEventListener('click', (e) => {
     e.preventDefault();
-    
+
     const title = document.querySelector('input[name="title"]').value.trim();
     const story = document.querySelector('textarea[name="story"]').value.trim();
     const firstName = document
@@ -27,14 +28,7 @@ export const createNewPost = () => {
 
     const newPost = { id, title, story, authorName, img };
 
-    let oldPosts = sessionStorage.getItem('newPost');
-    if (oldPosts) {
-      oldPosts = JSON.parse(oldPosts);
-      oldPosts.push(newPost);
-      sessionStorage.setItem('newPost', JSON.stringify(oldPosts));
-    } else {
-      sessionStorage.setItem('newPost', JSON.stringify([newPost]));
-    }
+    Storage.setNewPost(newPost);
 
     window.location.href = './home.html';
   });
